Add queue receive context abstraction
