+++ src/MyServiceBus/ReceiveContext.cs
@@ -6,20 +6,6 @@ using MyServiceBus.Serialization;
 
 namespace MyServiceBus;
 
-public interface ReceiveContext : PipeContext
-{
-    Guid MessageId { get; }
-    IList<string> MessageType { get; }
-    Uri? ResponseAddress { get; }
-    Uri? FaultAddress { get; }
-    Uri? ErrorAddress { get; }
-
-    IDictionary<string, object> Headers { get; }
-
-    bool TryGetMessage<T>([NotNullWhen(true)] out T? message)
-        where T : class;
-}
-
 public class ReceiveContextImpl : BasePipeContext, ReceiveContext
 {
     private readonly IMessageContext messageContext;
